pyproject.toml
[tool.poetry]
name = "psql"
version = "0.1.0"
description = ""
authors = ["user <panpanzov.nur@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
